Consolidate inventory and equipment panel
